+++ pkgs/dart_mcp/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Handle failed calls to `listRoots` in the `RootsTrackingSupport` mixin more
   gracefully. Previously this would leave the `roots` future hanging
   indefinitely, but now it will log an error and set the roots to empty.
+- Added validation for Schema extension.
 
 ## 0.2.0